requirements.txt
python-telegram-bot==20.7
yfinance==0.2.36
pandas==2.1.4
pandas-ta==0.3.14b
nsetools==1.0.11
bsedata==0.6.0
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
newsapi-python==0.2.7
pytz==2024.1
numpy==1.26.4
aiohttp==3.9.1
